requirements.txt
torch>=2.0.1
torchvision
torchaudio
tensorboard
numpy
matplotlib
seaborn
ipython
pytest
sox
scipy>=1.5.0
pandas
tqdm
pytorch-lightning
pyroomacoustics
PyYAML
rich
tabulate
librosa
hydra-core
hydra-joblib-launcher
onnxruntime
webdataset
soundfile
pesq
auraloss
torch_ema
pedalboard
gradio
fast_bss_eval
einops
jinja2
joblib
omegaconf
pystoi
setuptools
soxr
sympy
huggingface_hub
transformers
speechmos
python-Levenshtein
